requirements.txt
PyYAML~=6.0
mwclient~=0.10.1